clean up with getNonRoundedLocalRectangle
